data-process/stats-service/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
psycopg2-binary>=2.9
boto3>=1.35
numpy>=1.26
scipy>=1.13
python-dotenv>=1.0
